% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{From a given data frame drops selected columns and creates a new data frame}
\usage{
clean_data(data, columns)
}
\arguments{
\item{data}{A data frame or tibble}

\item{columns}{Names of columns to be dropped from the data frame}
}
\value{
A new data frame from the old data frame with selected columns dropped
}
\description{
From a given data frame drops selected columns and creates a new data frame
}
\examples{
clean_data(mtcars, c("mpg","cyl"))
}
