% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_boxplot.R
\name{create_boxplot}
\alias{create_boxplot}
\title{Create a ggplot boxplot}
\usage{
create_boxplot(data, x_var, y_var, title, x_title, y_title)
}
\arguments{
\item{data}{A data frame or tibble.}

\item{x_var}{Variable for the x axis of the plot}

\item{y_var}{Variable for the y axis of the plot}

\item{title}{Title of the plot}

\item{x_title}{Title of the x axis}

\item{y_title}{Title of the y axis}
}
\value{
A ggplot boxplot
}
\description{
A function for creating a side by side boxplot
}
\examples{
create_boxplot(mtcars, mtcars$cyl, mtcars$cyl, "MPG by No. Cylinders", "mpg", "No. Cylinders")
}
