% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{loads the data from the Internet and returns a data frame}
\usage{
load_data(url)
}
\arguments{
\item{url}{the path to the input file (a string of a URL)}
}
\value{
a data frame with the data from the given URL
}
\description{
loads the data from the Internet and returns a data frame
}
