% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_test_split.R
\name{train_test_split}
\alias{train_test_split}
\title{Splitting test data into testing and training data}
\usage{
train_test_split(data_frame, train_percent, col_name)
}
\arguments{
\item{data_frame}{A data frame or data frame extension (e.g. a tibble).}

\item{train_percent}{The percent of the parent dataframe to be train-test split}

\item{col_name}{The column name to group by in the testing dataframe}
}
\value{
A list with two data frames
A user-input percent of the parent dataframe training dataframe split of the parent dataframe
A user-input percent's compliment of the parent dataframe testing dataframe split of the parent dataframe
}
\description{
Creates two data frames, one for testing and one for training,
from the parent data frame
}
\examples{
train_test_split(mtcars, 0.75, "mpg")
}
